optimise name compression a bit
On a large message, this nearly halves the time and cuts allocations by
a factor of 3 (though the total volume of allocation is still almost as
large).

We could reduce the number of allocations further by storing
`recordedLabel`s contiguously instead of pointers, but this ups the
total bytes allocated and seems to slow performance a bit.

As a comparison point, writing a message with no name compression takes
about 1/3 the time and does virtually no allocation, so compression is
by far the most expensive part of writing a message.