pre-allocate 4 labels when parsing compressed names
This saves most re-allocating as we discover how long the name is.

Speeds up and reduces #allocations by about 20% for large and small
messages.

It does waste a bit of total allocation space, but counting the precise
number of names before allocating is not worth the complexity.

4 seems to be the optimal number on my machine, with go1.24 and the
particular sample messages I have.  Could definitely be a bit different
for messages with only longer or shorter names, or if a different go
version has `append` scale differently.