hook up our test dns server to the iterative resolver
This is now (kind of) a working recursive proxy!

I pointed my resolv.conf to this and restarted my browser with many tabs
open and it mostly worked, if a bit slowly.  A few things failed to
load, perhaps because of timeouts, or perhaps because I'm still
returning gibberish for AAAA requests...

There's *a lot* of dns traffic happening in the background on a modern
mac!